Cast the last param to DWORD to avoid a compiler warning

+++ OpenKeys/OpenKeys.cpp
@@ -369,7 +369,7 @@ void AddToStartup() {
     std::wstring progPath = GetExecutableDirectory() + L"\\OpenKeys.exe";
     HKEY hkey = NULL;
     LONG createStatus = RegCreateKey(HKEY_CURRENT_USER, L"SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Run", &hkey); //Creates a key       
-    LONG status = RegSetValueEx(hkey, L"MyApp", 0, REG_SZ, (BYTE*)progPath.c_str(), (progPath.size() + 1) * sizeof(wchar_t));
+    LONG status = RegSetValueEx(hkey, L"MyApp", 0, REG_SZ, (BYTE*)progPath.c_str(), static_cast<DWORD>((progPath.size() + 1) * sizeof(wchar_t)));
 }
 
 void LoadShortcuts() {
